docs: document test suites in README and align e2e test naming
- Add a "Test Suites at a Glance" table to README describing the scope
  of unit, integration, e2e, load, and HA end-to-end test suites
- Rename e2e *Test classes to *IT so they run under the integration
  profile (Failsafe) instead of the unit phase
- Adjust mvn-test workflow accordingly
