+++ e2e/src/test/java/com/arcadedb/e2e/JsFunctionsIT.java
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class JsFunctionsTest extends ArcadeContainerTemplate {
+class JsFunctionsIT extends ArcadeContainerTemplate {
   private RemoteDatabase database;
 
   @BeforeEach
