+++ e2e/src/test/java/com/arcadedb/e2e/RemoteGrpcDatabaseIT.java
@@ -19,20 +19,26 @@
 package com.arcadedb.e2e;
 
 import com.arcadedb.query.sql.executor.ResultSet;
-import com.arcadedb.remote.RemoteDatabase;
+import com.arcadedb.remote.grpc.RemoteGrpcDatabase;
+import com.arcadedb.remote.grpc.RemoteGrpcServer;
 import com.arcadedb.utility.CollectionUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-class RemoteDatabaseQueriesTest extends ArcadeContainerTemplate {
-  private RemoteDatabase database;
+class RemoteGrpcDatabaseIT extends ArcadeContainerTemplate {
+
+  private RemoteGrpcDatabase database;
 
   @BeforeEach
   void setUp() {
-    database = new RemoteDatabase(host, httpPort, "beer", "root", "playwithdata");
+
+    RemoteGrpcServer server = new RemoteGrpcServer(host, grpcPort, "root", "playwithdata", true, List.of());
+    database = new RemoteGrpcDatabase(server, host, grpcPort, httpPort, "beer", "root", "playwithdata");
     // ENLARGE THE TIMEOUT TO PASS THESE TESTS ON CI (GITHUB ACTIONS)
     database.setTimeout(60_000);
   }
@@ -45,26 +51,20 @@ class RemoteDatabaseQueriesTest extends ArcadeContainerTemplate {
 
   @Test
   void simpleSQLQuery() {
-    database.transaction(() -> {
-      final ResultSet result = database.query("SQL", "select * from Beer limit 10");
-      assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
-    }, true, 10);
+    final ResultSet result = database.query("SQL", "select * from Beer limit 10");
+    assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
   }
 
   @Test
   void simpleGremlinQuery() {
-    database.transaction(() -> {
-      final ResultSet result = database.query("gremlin", "g.V().limit(10)");
-      assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
-    }, false, 10);
+    final ResultSet result = database.query("gremlin", "g.V().limit(10)");
+    assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
   }
 
   @Test
   void simpleCypherQuery() {
-    database.transaction(() -> {
-      final ResultSet result = database.query("cypher", "MATCH(p:Beer) RETURN * LIMIT 10");
-      assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
-    }, false, 10);
+    final ResultSet result = database.query("cypher", "MATCH(p:Beer) RETURN * LIMIT 10");
+    assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
   }
 
   @Test
@@ -74,4 +74,17 @@ class RemoteDatabaseQueriesTest extends ArcadeContainerTemplate {
       assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
     }, false, 10);
   }
+
+  @Test
+  void streamQueryWithSQL() {
+    final ResultSet result = database.queryStream("sql", "select * from Beer limit 10");
+    assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
+  }
+
+  @Test
+  void streamQueryWithOpenCypher() {
+    final ResultSet result = database.queryStream("opencypher", "MATCH(p:Beer) RETURN p LIMIT 10");
+    assertThat(CollectionUtils.countEntries(result)).isEqualTo(10);
+  }
+
 }
